+++ demos/chinook/src/main/java/is/codion/framework/demos/chinook/tutorial/MultiPanelDemo.java
@@ -56,7 +56,7 @@ public final class MultiPanelDemo {
 		JPanel basePanel = new JPanel(gridLayout(2, 2));
 		for (int i = 0; i < 4; i++) {
 			EntityPanel employeePanel = new EntityPanel(employeeModel, new EmployeeEditPanel(employeeModel.editModel()));
-			employeePanel.tablePanel().conditions().conditionView().set(SIMPLE);
+			employeePanel.tablePanel().conditions().view().set(SIMPLE);
 			employeePanel.initialize();
 			basePanel.add(employeePanel);
 		}